+++ src/components/Form/DragnDropImage/index.tsx
@@ -1,9 +1,9 @@
 import { CircularProgress } from '@material-ui/core';
 import { Text, View } from 'components/Common';
 import { LineAwesomeIcon } from 'components/Icons';
-import React, { ChangeEvent, DragEvent, FC } from 'react';
-import { buildStyles, ClassNameProps, colors, mx, StyleProps } from 'styles';
-
+import React, { ChangeEvent, DragEvent, FC, useState } from 'react';
+import { buildStyles, ClassNameProps, colors, ms, mx, StyleProps } from 'styles';
+import { useDropzone } from 'react-dropzone';
 import { FormSelectFileBtn } from '../SelectFileBtn';
 
 interface Props extends StyleProps, ClassNameProps {
@@ -13,6 +13,10 @@ interface Props extends StyleProps, ClassNameProps {
   border?: boolean;
   fullWidth?: boolean;
   onFileSelect?: (file: File) => void;
+}
+
+interface DragImageFile extends File {
+  preview?: string;
 }
 
 export const FormDragnDropImage: FC<Props> = ({
@@ -25,6 +29,8 @@ export const FormDragnDropImage: FC<Props> = ({
   disabled,
   onFileSelect,
 }) => {
+  const [files, setFiles] = useState<DragImageFile[]>([]);
+
   const handleClick = (e: DragEvent<HTMLInputElement>) => {
     e.preventDefault();
     return false;
@@ -42,46 +48,74 @@ export const FormDragnDropImage: FC<Props> = ({
     }
   };
 
+  const { getRootProps, getInputProps } = useDropzone({
+    accept: 'image/*',
+    onDrop: acceptedFiles => {
+      setFiles(
+        acceptedFiles.map(file =>
+          Object.assign(file, {
+            preview: URL.createObjectURL(file),
+          }),
+        ),
+      );
+      acceptedFiles.map(file => onFileSelect && onFileSelect(file));
+    },
+  });
+
+  const getPreviewSrc = () => {
+    if (src) {
+      return src;
+    }
+    if (files.length) {
+      return files[0].preview;
+    }
+    return undefined;
+  };
+
+  const previewSrc = getPreviewSrc();
+
   const styles = useStyles();
 
   return (
-    <View
-      className={className}
-      style={[
-        styles.container,
-        fullWidth && mx.w100,
-        !!src && { backgroundImage: `url(${src})` },
-        border && !src && mx.border(3, 'dashed', colors.silverTwo),
-        style,
-      ]}
-      row
-    >
-      <input style={styles.input} onClick={handleClick} type="file" onChange={handleChange} />
-      {!processing && (
-        <View style={styles.content} row alignItems="center">
-          {!src && (
-            <View column style={styles.iconWrap}>
-              <LineAwesomeIcon
-                style={styles.icon}
-                size={62}
-                type="image"
-                color={!disabled ? colors.coolBlueTwo : colors.veryLightPinkTwo}
-              />
-              <Text style={[styles.iconText, disabled && { color: colors.veryLightPinkTwo }]} block>
-                {'Drag an image here'}
-              </Text>
-            </View>
-          )}
-          {!src && <Text style={[styles.orText, disabled && { color: colors.veryLightPinkTwo }]}>{'Or'}</Text>}
-          <FormSelectFileBtn style={styles.btnWrap} disabled={disabled} btnStyle={styles.btn} onFileSelect={onFileSelect} />
-        </View>
-      )}
-      {processing && (
-        <View style={styles.content} alignItems="center" justifyContent="center">
-          <CircularProgress />
-        </View>
-      )}
-    </View>
+    <div {...getRootProps()}>
+      <input {...getInputProps()} style={styles.input} onClick={handleClick} type="file" onChange={handleChange} />
+      <View
+        className={className}
+        style={ms(
+          styles.container,
+          style,
+          { backgroundImage: previewSrc ? `url(${previewSrc})` : undefined },
+          border && !previewSrc && styles.containerBorder,
+          fullWidth && mx.w100,
+        )}
+        row
+      >
+        {!processing && (
+          <View style={styles.content} row alignItems="center">
+            {!previewSrc && (
+              <View column style={styles.iconWrap}>
+                <LineAwesomeIcon
+                  style={styles.icon}
+                  size={62}
+                  type="image"
+                  color={!disabled ? colors.coolBlueTwo : colors.veryLightPinkTwo}
+                />
+                <Text style={[styles.iconText, disabled && { color: colors.veryLightPinkTwo }]} block>
+                  {'Drag an image here'}
+                </Text>
+              </View>
+            )}
+            {!previewSrc && <Text style={[styles.orText, disabled && { color: colors.veryLightPinkTwo }]}>{'Or'}</Text>}
+            <FormSelectFileBtn style={styles.btnWrap} disabled={disabled} btnStyle={styles.btn} onFileSelect={onFileSelect} />
+          </View>
+        )}
+        {processing && (
+          <View style={styles.content} alignItems="center" justifyContent="center">
+            <CircularProgress />
+          </View>
+        )}
+      </View>
+    </div>
   );
 };
 
@@ -99,6 +133,9 @@ const useStyles = buildStyles(({ isMobile, whenMobile }) => ({
     backgroundPosition: 'center',
     backgroundRepeat: 'no-repeat',
     boxSizing: 'border-box',
+  },
+  containerBorder: {
+    ...mx.border(3, 'dashed', colors.silverTwo),
   },
   input: {
     position: 'absolute',